using bower now
ok bower adds alot of stuff to the repo but still it makes everything a
little better to use. i like how easy it is to upgrade packages and
manage them. maybe i will find a better way of handling them but till
now its the way i use it.