use vagrant synced folder instead of cloned repo
i used to clone the repo again into the vagrant box under
`/home/vagrant/genesis`. this is kind of boilerplate with vagrant because
it mounts the vagrant dir (aka the repo) under `/vagrant`. the only
problem is if your user running vagrant has a different uid. so its some
fiddling with `usermod -u 1001 vagrant` in vagrant. but it works.