reworked vagrant provision for windows hosts
there is a problem with running `vagrant up` under windows because
ansible cant be installed. i use a special shell script running the
ansible playbook on the virtual machine itself. there was alot of
problems because virtualbox shared folders set the execution bit on all
files and ansible thinks that are scripts. i needed some hacks to fix
that. alot of copying files to /tmp and stuff. but it should work now.